feat: group list 컴포넌트 분리 및 tablist onChange 를 optional로 변경
